remove junk files, move datasets_controller
