--- stash/stash_api/app/assets/javascripts/stash_api/files.js
@@ -1,2 +1,0 @@
-// Place all the behaviors and hooks related to the matching controller here.
-// All this logic will automatically be available in application.js.
