--- stash/stash_api/app/assets/stylesheets/stash_api/files.css
@@ -1,4 +1,0 @@
-/*
-  Place all the styles related to the matching controller here.
-  They will automatically be included in application.css.
-*/
