--- stash/stash_api/app/helpers/stash_api/application_helper.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module StashApi
-  module ApplicationHelper
-  end
-end
