--- stash/stash_api/app/helpers/stash_api/files_helper.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module StashApi
-  module FilesHelper
-  end
-end
