--- stash/stash_api/app/helpers/stash_api/users_helper.rb
@@ -1,4 +1,0 @@
-module StashApi
-  module UsersHelper
-  end
-end
